+++ telemonitor/helpers.py
@@ -152,7 +152,7 @@ def cli_arguments_parser() -> object:
 
     adv_group = argparser.add_argument_group('advanced optional arguments')
     adv_group.add_argument('--dev', help='enable unstable development features', action='store_true', dest='dev_features')
-    adv_group.add_argument('--verbose', '-v', help='write more detailed information to log file', action='store_true')
+    adv_group.add_argument('--verbose', '-v', help='write debug information to log file', action='store_true')
     adv_group.add_argument('--config-check', action='store_true', help='run config file initialization procedure and exit', dest='config_check_only')
     adv_group.add_argument('--no-config-check', action='store_true', help="don't scan configuration file on start", dest='disable_config_check')